refactor: fix bug and refactor
