+++ src/main/kotlin/net/ccbluex/liquidbounce/web/theme/component/types/minimap/ChunkRenderer.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
  *
- * Copyright (c) 2015 - 2024 CCBlueX
+ * Copyright (c) 2015-2024 CCBlueX
  *
  * LiquidBounce is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -15,8 +15,10 @@
  *
  * You should have received a copy of the GNU General Public License
  * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ *
+ *
  */
-package net.ccbluex.liquidbounce.features.module.modules.render.minimap
+package net.ccbluex.liquidbounce.web.theme.component.types.minimap
 
 import net.ccbluex.liquidbounce.utils.block.ChunkScanner
 import net.ccbluex.liquidbounce.utils.client.mc
@@ -39,7 +41,7 @@ object ChunkRenderer {
 
     fun unloadEverything() {
         heightmapManager.unloadAllChunks()
-        this.textureAtlasManager.deallocateAll()
+        textureAtlasManager.deallocateAll()
     }
 
     fun getAtlasPosition(chunkPos: ChunkPos): MinimapTextureAtlasManager.AtlasPosition {
@@ -47,7 +49,7 @@ object ChunkRenderer {
     }
 
     fun prepareRendering(): Int {
-        return this.textureAtlasManager.prepareRendering()
+        return textureAtlasManager.prepareRendering()
     }
 
     object MinimapChunkUpdateSubscriber : ChunkScanner.BlockChangeSubscriber {
